% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline}
\alias{baseline}
\title{Baseline Scenario}
\usage{
baseline(df, dep)
}
\arguments{
\item{df}{The dataframe you wish to analyze}

\item{dep}{The dependent variable}
}
\value{
b1, p-value and confidence interval for baseline scenario
}
\description{
Baseline Scenario
}
