% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice.R
\name{choice}
\alias{choice}
\title{Choice analysis, chooses a significant p-value from deviations if baseline is insignificant}
\usage{
choice(scenarios)
}
\arguments{
\item{scenarios}{The dataframe you wish to analyze, typically a dataframe from the analysis function}
}
\value{
b1, p-value and confidence interval for chosen scenario per iteration. As well as how many deviations were significant.
}
\description{
Choice analysis, chooses a significant p-value from deviations if baseline is insignificant
}
