% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{conditions}
\alias{conditions}
\title{Combines the different deviation conditions and baseline condition}
\usage{
conditions(df, dep)
}
\arguments{
\item{df}{The dataframe you wish to analyze}

\item{dep}{The dependent variable}
}
\value{
b1, p-value and confidence interval for all scenarios
}
\description{
Combines the different deviation conditions and baseline condition
}
