% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{dgm}
\alias{dgm}
\title{Data generating mechanism}
\usage{
dgm(n, b0, b1, b_z, b_d)
}
\arguments{
\item{n}{Sample size in the baseline scenario}

\item{b0}{Intercept value}

\item{b1}{Main effect coefficient in the effect scenario}

\item{b_z}{Continuous covariate effect coefficient}

\item{b_d}{Dichotomous covariate effect coefficient}
}
\value{
Dataframe
}
\description{
Data generating mechanism
}
