% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr.R
\name{extr}
\alias{extr}
\title{Extracts and calculates coefficients, p-value and confidence interval of a
model}
\usage{
extr(model)
}
\arguments{
\item{model}{The lm model that you wish to extract input from}
}
\value{
b1, p-value and confidence interval of input model
}
\description{
Extracts and calculates coefficients, p-value and confidence interval of a
model
}
