% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{models}
\title{Adds a continuous or dichotomous covariate to the linear regression, or uses
alternative outcome.}
\usage{
models(df, dep)
}
\arguments{
\item{df}{The dataframe you wish to analyze}

\item{dep}{The dependent variable}
}
\value{
b1, p-value and confidence interval for the model deviations
}
\description{
Adds a continuous or dichotomous covariate to the linear regression, or uses
alternative outcome.
}
