% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsig.R
\name{nsig}
\alias{nsig}
\title{Significance counter}
\usage{
nsig(df)
}
\arguments{
\item{df}{Input a dataframe which follows the output structure of the analysis function}
}
\value{
Dataframe with the number of significant p-values per condition
}
\description{
Significance counter
}
