% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{outlier}
\alias{outlier}
\title{Removes outliers based on z-score > 2 and Cook's distance > 1}
\usage{
outlier(df, dep)
}
\arguments{
\item{df}{The dataframe you wish to analyze}

\item{dep}{The dependent variable}
}
\value{
b1, p-value and confidence interval for the outlier deviations
}
\description{
Removes outliers based on z-score > 2 and Cook's distance > 1
}
