% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize.R
\name{samplesize}
\alias{samplesize}
\title{Changes your sample size with 2.5\%/15\% increase and decrease.}
\usage{
samplesize(df, dep)
}
\arguments{
\item{df}{The dataframe you wish to analyze}

\item{dep}{The dependent variable}
}
\value{
b1, p-value and confidence interval for sample size deviations
}
\description{
Changes your sample size with 2.5\%/15\% increase and decrease.
}
