% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\title{Overarching function (generates data, runs each condition, extracts stats)}
\usage{
analysis(iter, n, b0, b1, b_z, b_d)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Sample size in baseline scenario}

\item{b0}{Intercept value}

\item{b1}{Main effect coefficient in the effect scenario}

\item{b_z}{Continuous covariate effect coefficient}

\item{b_d}{Dichotomous covariate effect coefficient}
}
\value{
b1, p-value and confidence interval for baseline and each deviation
scenario
}
\description{
Overarching function (generates data, runs each condition, extracts stats)
}
